Corrected mistakes and made the code easier to use
GuestRequest:
I corrected mistakes in ToString and made the constructor easier to use.
Host:
I corrected mistakes in its constructor and made it IEnumerabled.
HostingUnit;
I deleted the function 'rest' - because we actually don't need it - and fixed its 'ToString'. Generally I ofer to add '\n's to its ToString.
Program:
I copied some of the code from the pdf file, and started the function 'CreateRandomRequest' for tests but you still have to correct and finish it